refactor: move accounting_init primitives to module
